backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Maksym Potapchuk",email = "maksym.potapchuk.hp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-filter (>=25.0.0,<26.0.0)",
    "ruff (>=0.15.1,<0.16.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


# ── Ruff ─────────────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 88

# ── Formatter ─────────────────────────────────────────────────────────────────
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

# ── Linter ────────────────────────────────────────────────────────────────────
[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade — modernise syntax automatically
    "B",    # flake8-bugbear
    "C90",  # mccabe complexity
    "DJ",   # flake8-django  ← Django/DRF patterns
    "SIM",  # flake8-simplify
    "PIE",  # flake8-pie
    "PERF", # perflint — performance anti-patterns
    "LOG",  # flake8-logging
    "TCH",  # flake8-type-checking — move pure-type imports to TYPE_CHECKING
    "TID",  # flake8-tidy-imports — enforce absolute imports
    "RUF",  # ruff-native rules
]
ignore = [
    "E501",   # line too long — the formatter handles this
    "B008",   # function call in default arg — needed for DRF (permission_classes=[...])
    "DJ001",  # null=True on string fields — intentional in several models
    "SIM108", # ternary operator — sometimes hurts readability
    "TID252", # relative imports from parent — allowed within the same Django app
]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*.py" = ["ALL"]
"*/settings*.py"    = ["E501", "F401", "F403", "F405"]
"*/tests*.py"       = ["S101", "B011"]
"conftest.py"       = ["S101"]
# DRF / Django class attributes are intentionally mutable — suppress false positives
"*/views.py"        = ["RUF012"]
"*/serializers.py"  = ["RUF012"]
"*/models.py"       = ["RUF012"]  # Django Meta inner-class attrs (ordering, constraints…)
"*/filters.py"      = ["RUF012"]  # django-filter FilterSet Meta attrs

# ── isort ─────────────────────────────────────────────────────────────────────
[tool.ruff.lint.isort]
known-first-party  = ["account", "driver", "vehicle", "fleet_management", "config"]
known-third-party  = ["django", "rest_framework", "corsheaders", "django_filters", "jwt"]
force-sort-within-sections = true
combine-as-imports = true

# ── Complexity ────────────────────────────────────────────────────────────────
[tool.ruff.lint.mccabe]
max-complexity = 10

# ── Naming ────────────────────────────────────────────────────────────────────
[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
    "classmethod",
    "django.utils.decorators.classonlymethod",
]
